Fixes poppy stuff, makes it so we can actually build it :D
