+++ Code/Monos/SkillPointMono.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+
+namespace PoppyScyyeGameModes.Monos
+{
+    public class SkillPointMono : MonoBehaviour
+    {
+        public void Start()
+        {
+            // SkillPointShop.SkillPoints.Add("SkillPoints", 10);
+        }
+    }
+}
