+++ Code/PoppyScyyeGameModes.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34031.279
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PoppyScyyeGameModes", "PoppyScyyeGameModes.csproj", "{2C730035-9EB7-4BE1-B0DC-F88F58916BEA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2C730035-9EB7-4BE1-B0DC-F88F58916BEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2C730035-9EB7-4BE1-B0DC-F88F58916BEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2C730035-9EB7-4BE1-B0DC-F88F58916BEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2C730035-9EB7-4BE1-B0DC-F88F58916BEA}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {873B276D-B034-47C8-9B21-6753EDFEB143}
+	EndGlobalSection
+EndGlobal
