--- Code/PoppysAndScyyesGamemodes/ExtensionsLMAO/CharacterStatModifiersExtension.cs
@@ -1,44 +1,0 @@
-﻿using System;
-using System.Runtime.CompilerServices;
-using HarmonyLib;
-
-namespace PSGM.Extensions
-{
-    // ADD FIELDS TO GUN
-    [Serializable]
-    public class CharacterStatModifiersAdditionalData
-    {
-
-        //public int damageSkillPoints;
-        public CharacterStatModifiersAdditionalData()
-        {
-           // damageSkillPoints = 0;
-        }
-    }
-    public static class CharacterStatModifiersExtension
-    {
-        public static readonly ConditionalWeakTable<CharacterStatModifiers, CharacterStatModifiersAdditionalData> data = new ConditionalWeakTable<CharacterStatModifiers, CharacterStatModifiersAdditionalData>();
-
-        public static CharacterStatModifiersAdditionalData GetAdditionalData(this CharacterStatModifiers statModifiers)
-        {
-            return data.GetOrCreateValue(statModifiers);
-        }
-
-        public static void AddData(this CharacterStatModifiers statModifiers, CharacterStatModifiersAdditionalData value)
-        {
-            try
-            {
-                data.Add(statModifiers, value);
-            }
-            catch (Exception) { }
-        }
-    }
-    [HarmonyPatch(typeof(CharacterStatModifiers), "ResetStats")]
-    class CharacterStatModifiersPatchResetStats
-    {
-        private static void Prefix(CharacterStatModifiers __instance)
-        {
-            //__instance.GetAdditionalData().damageSkillPoints = 0;
-        }
-    }
-}
